Created RefactorizationSolver class.
